requirements.txt
#
# UNOC - requirements.txt
#

Flask==3.0.3
PyYAML==6.0.1
Flask-Cors==4.0.1
pydantic==2.8.2
pytest==8.2.2
